+++ src/views/Dashboard/Cards/StockBalanceTile.tsx
@@ -1,0 +1,29 @@
+import * as React from 'react';
+import { PriceTag } from '../../../components/PriceTag';
+import { StatusCard } from '../../../components/StatusCard/StatusCard';
+
+interface StockBalanceTileProps {
+    value: number;
+}
+
+interface StockBalanceTileState {
+}
+
+class StockBalanceTile extends React.Component<StockBalanceTileProps, StockBalanceTileState> {
+
+    constructor( props: StockBalanceTileProps ) {
+        super( props );
+    }
+
+    render() {
+        return (
+            <StatusCard
+                icon={<i className="pe-7s-graph2 text-warning"/>}
+                title="Stock Balance"
+                value={<PriceTag value={this.props.value}/>}
+                noFooter={true}
+            />
+        );
+    }
+}
+export default StockBalanceTile;
