+++ src/views/Dashboard/Cards/index.ts
@@ -1,0 +1,4 @@
+import BalanceTile from './BalanceTile';
+import StockBalanceTile from './StockBalanceTile';
+
+export { BalanceTile, StockBalanceTile };
